Insert de-duping code into find-walks
In some cases of ambiguity (e.g. S = S;'a'.) there may be more than
one Earley item that covers the same range for a given
non-terminal. (Here, for the input 'a', S will have one item covering 0
to 1 using 'a' and another using S.)  This was leading to apparent
duplication of rules in the parse-forest grammar.